app/requirements.txt
fastapi
pydantic
uvicorn
python-multipart
docker
numpy
tritonclient[http]
pillow
pandas
scipy
opencv-python